streamlit_app/requirements.txt
streamlit
requests
pandas
plotly
python-dotenv
pytest
pytest-mock
